lesson02-ml-pipeline-mlflow/pyproject.toml
[project]
name = "lesson02-ml-pipeline-mlflow"
version = "0.1.0"
description = "Lesson 02 ML pipeline and MLflow mini-application."
requires-python = ">=3.13,<3.14"
dependencies = [
    "pandas",
    "mlflow",
    "scikit-learn",
    "numpy",
    "agent-framework-azure-ai==1.0.0rc3",
    "agent-framework-core==1.0.0rc3",
    "azure-identity",
    "python-dotenv",
]

[tool.uv]
package = false